+++ HISTORY.rst
@@ -12,6 +12,9 @@ Unreleased
 
 **Improvements**
 
+- create_or_update lyrics accepts now a model so we can change its env (user,
+  context, ...)
+
 **Documentation**
 
 **Build**